requirements.txt
adal==1.2.7
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
idna==3.3
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lxml==4.7.1
MarkupSafe==2.0.1
mistune==0.8.4
msal==1.16.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.3
ntlm-auth==1.5.0
office365-rest-client==2.2.0
Office365-REST-Python-Client==2.3.9
packaging==21.3
pandocfilters==1.4.3
parso
pexpect==4.8.0
pickleshare==0.7.5
pip==21.2.2
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.4
pyrsistent==0.17.3
python-dateutil==2.8.2
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.10.0
requests==2.27.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
Send2Trash==1.8.0
setuptools==58.0.4
SharePlum==0.5.1
six==1.16.0
terminado==0.9.4
testpath==0.5.0
tornado==6.1
traitlets==4.3.3
urllib3==1.26.8
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.0
widgetsnbextension==3.5.1
